Add 2 new recipes and add width attribute to lasagna images
Added a new recipe page for pb_mousse with content and image, and
a new recipe page for banana_bread with content and image. To keep
image size consistent, the width attribute was added to the new
recipe pages and updated the lasanga page to add this to the already
existing image for consistency in presentation.